requirements.txt
torch
torchvision
tqdm
transformers
gradio
numpy
pillow
uvicorn
fastapi
starlette
scikit-learn